+++ data-prepper-plugins/lambda/src/integrationTest/java/org/opensearch/dataprepper/plugins/sink/lambda/LambdaSinkServiceIT.java
@@ -30,12 +30,14 @@ import org.opensearch.dataprepper.model.plugin.PluginFactory;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.model.sink.OutputCodecContext;
 import org.opensearch.dataprepper.model.types.ByteCount;
-import org.opensearch.dataprepper.plugins.common.lambda.accumlator.BufferFactory;
-import org.opensearch.dataprepper.plugins.common.lambda.accumlator.InMemoryBufferFactory;
-import org.opensearch.dataprepper.plugins.common.lambda.config.AwsAuthenticationOptions;
-import org.opensearch.dataprepper.plugins.common.lambda.config.BatchOptions;
-import org.opensearch.dataprepper.plugins.common.lambda.config.ThresholdOptions;
-import org.opensearch.dataprepper.plugins.sink.lambda.dlq.DlqPushHandler;
+import org.opensearch.dataprepper.plugins.lambda.common.accumlator.BufferFactory;
+import org.opensearch.dataprepper.plugins.lambda.common.accumlator.InMemoryBufferFactory;
+import org.opensearch.dataprepper.plugins.lambda.common.config.AwsAuthenticationOptions;
+import org.opensearch.dataprepper.plugins.lambda.common.config.BatchOptions;
+import org.opensearch.dataprepper.plugins.lambda.common.config.ThresholdOptions;
+import org.opensearch.dataprepper.plugins.lambda.sink.LambdaSinkConfig;
+import org.opensearch.dataprepper.plugins.lambda.sink.LambdaSinkService;
+import org.opensearch.dataprepper.plugins.lambda.sink.dlq.DlqPushHandler;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.lambda.LambdaClient;
