--- data-prepper-plugins/lambda/src/test/java/org/opensearch/dataprepper/plugins/sink/lambda/LambdaClientFactoryTest.java
@@ -1,97 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-package org.opensearch.dataprepper.plugins.sink.lambda;
-
-import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.ArgumentCaptor;
-import static org.mockito.ArgumentMatchers.any;
-import org.mockito.Mock;
-import org.mockito.MockedStatic;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.opensearch.dataprepper.aws.api.AwsCredentialsOptions;
-import org.opensearch.dataprepper.aws.api.AwsCredentialsSupplier;
-import org.opensearch.dataprepper.plugins.common.lambda.config.AwsAuthenticationOptions;
-import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.lambda.LambdaClient;
-import software.amazon.awssdk.services.lambda.LambdaClientBuilder;
-
-import java.util.Map;
-import java.util.UUID;
-
-@ExtendWith(MockitoExtension.class)
-class LambdaClientFactoryTest {
-    @Mock
-    private LambdaSinkConfig lambdaSinkConfig;
-    @Mock
-    private AwsCredentialsSupplier awsCredentialsSupplier;
-
-    @Mock
-    private AwsAuthenticationOptions awsAuthenticationOptions;
-
-    @BeforeEach
-    void setUp() {
-        when(lambdaSinkConfig.getAwsAuthenticationOptions()).thenReturn(awsAuthenticationOptions);
-    }
-
-    @Test
-    void createLambdaClient_with_real_LambdaClient() {
-        when(awsAuthenticationOptions.getAwsRegion()).thenReturn(Region.US_EAST_1);
-        final LambdaClient lambdaClient = LambdaClientFactory.createLambdaClient(lambdaSinkConfig, awsCredentialsSupplier);
-
-        assertThat(lambdaClient, notNullValue());
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"us-east-1", "us-west-2", "eu-central-1"})
-    void createlambdaClient_provides_correct_inputs(final String regionString) {
-        final Region region = Region.of(regionString);
-        final String stsRoleArn = UUID.randomUUID().toString();
-        final Map<String, String> stsHeaderOverrides = Map.of(UUID.randomUUID().toString(), UUID.randomUUID().toString());
-        when(awsAuthenticationOptions.getAwsRegion()).thenReturn(region);
-        when(awsAuthenticationOptions.getAwsStsRoleArn()).thenReturn(stsRoleArn);
-        when(awsAuthenticationOptions.getAwsStsHeaderOverrides()).thenReturn(stsHeaderOverrides);
-
-        final AwsCredentialsProvider expectedCredentialsProvider = mock(AwsCredentialsProvider.class);
-        when(awsCredentialsSupplier.getProvider(any())).thenReturn(expectedCredentialsProvider);
-
-        final LambdaClientBuilder lambdaClientBuilder = mock(LambdaClientBuilder.class);
-        when(lambdaClientBuilder.region(region)).thenReturn(lambdaClientBuilder);
-        when(lambdaClientBuilder.credentialsProvider(any())).thenReturn(lambdaClientBuilder);
-        when(lambdaClientBuilder.overrideConfiguration(any(ClientOverrideConfiguration.class))).thenReturn(lambdaClientBuilder);
-        try(final MockedStatic<LambdaClient> lambdaClientMockedStatic = mockStatic(LambdaClient.class)) {
-            lambdaClientMockedStatic.when(LambdaClient::builder)
-                    .thenReturn(lambdaClientBuilder);
-            LambdaClientFactory.createLambdaClient(lambdaSinkConfig, awsCredentialsSupplier);
-        }
-
-        final ArgumentCaptor<AwsCredentialsProvider> credentialsProviderArgumentCaptor = ArgumentCaptor.forClass(AwsCredentialsProvider.class);
-        verify(lambdaClientBuilder).credentialsProvider(credentialsProviderArgumentCaptor.capture());
-
-        final AwsCredentialsProvider actualCredentialsProvider = credentialsProviderArgumentCaptor.getValue();
-
-        assertThat(actualCredentialsProvider, equalTo(expectedCredentialsProvider));
-
-        final ArgumentCaptor<AwsCredentialsOptions> optionsArgumentCaptor = ArgumentCaptor.forClass(AwsCredentialsOptions.class);
-        verify(awsCredentialsSupplier).getProvider(optionsArgumentCaptor.capture());
-
-        final AwsCredentialsOptions actualCredentialsOptions = optionsArgumentCaptor.getValue();
-        assertThat(actualCredentialsOptions.getRegion(), equalTo(region));
-        assertThat(actualCredentialsOptions.getStsRoleArn(), equalTo(stsRoleArn));
-        assertThat(actualCredentialsOptions.getStsHeaderOverrides(), equalTo(stsHeaderOverrides));
-    }
-}
